+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add a Deed of novation and variation task to Transfer projects
 - Add Articles of Association task to Transfer projects
 - Add commercial transfer agreement task to Transfer projects
+- Add All conditions met? task to Transfer projects
 
 ### Changed